% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_geom_punto.R
\name{a_geom_punto}
\alias{a_geom_punto}
\title{A geom that writes points but is called punto}
\usage{
a_geom_punto(...)
}
\arguments{
\item{...}{all the geom_point arguments}
}
\value{

}
\description{
A geom that writes points but is called punto
}
\examples{
# without the function
library(ggplot2)
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 geom_point()

# the function
a_geom_punto
a_geom_punto

# using the function
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 a_geom_punto()

}
