% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_geom_point_blue.R
\name{b_geom_point_blue}
\alias{b_geom_point_blue}
\title{A geom that writes points and default color is blue}
\usage{
b_geom_point_blue(...)
}
\arguments{
\item{...}{all the geom_point arguments}
}
\value{

}
\description{
A geom that writes points and default color is blue
}
\examples{
# without the function
library(ggplot2)
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 geom_point(color = "blue")

# the function
b_geom_point_blue

# using the function
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 b_geom_point_blue()

}
