% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_geom_lm.R
\name{c_geom_lm}
\alias{c_geom_lm}
\title{A geom that draws the simple linear ols model}
\usage{
c_geom_lm(...)
}
\arguments{
\item{...}{all the geom_point arguments}
}
\value{

}
\description{
A geom that draws the simple linear ols model
}
\examples{
# without the function
library(ggplot2)
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 geom_point() +
 geom_smooth(method = lm, se = FALSE)

# the function
c_geom_lm

# using the function
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 geom_point() +
 c_geom_lm()

}
