% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_geom_segment_min_max.R
\name{d_geom_segment_mins_maxs}
\alias{d_geom_segment_mins_maxs}
\title{A geom that draws a segment from the minimum x and y to maximum x and y}
\usage{
d_geom_segment_mins_maxs(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{all the geom_point arguments}
}
\value{

}
\description{
A geom that draws a segment from the minimum x and y to maximum x and y
}
\examples{
# without the function
library(magrittr)
library(dplyr)
cars \%>\%
 summarize(
   min_speed = min(speed),
   max_speed = max(speed),
   min_dist = min(dist),
   max_dist = max(dist)
 ) ->
 mins_maxs

library(ggplot2)
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 geom_point() +
 geom_segment(data = mins_maxs,
     aes(x = min_speed, xend = max_speed,
         y = min_dist, yend = max_dist))
 layer_data(last_plot(), i = 2)

# the proto
"StatSegmentminmax"
d_geom_segment_mins_maxs

# using the function
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 geom_point() +
 d_geom_segment_mins_maxs()
 layer_data(last_plot(), i = 2)

}
