% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_geom_lm_intercept.R
\name{e_geom_lm_intercept}
\alias{e_geom_lm_intercept}
\title{A geom that draws a point at the intercept of linear model estimate}
\usage{
e_geom_lm_intercept(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{all the geom_point arguments}
}
\value{

}
\description{
A geom that draws a point at the intercept of linear model estimate
}
\examples{
# without the function
library(magrittr)
library(dplyr)
cars \%>\%
 lm(dist ~ speed, data = .) ->
 cars_model

 tibble(y_val = cars_model$coefficients[1],
 x_val = 0) ->
 cars_model_intercept

library(ggplot2)
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 geom_point() +
 geom_smooth(method = lm) +
 geom_point(data = cars_model_intercept,
     aes(x = x_val,
         y = y_val),
         color = "red",
         size = 3)

# the proto and function
"StatOlsintercept"
e_geom_lm_intercept

# using the function
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 geom_point() +
 geom_smooth(method = lm) +
 e_geom_lm_intercept(size = 4) +
 aes(color = speed > 12)

}
