% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_geom_line_endpoint.R
\name{f_geom_line_endpoint}
\alias{f_geom_line_endpoint}
\title{A geom that draws a point at the intercept of linear model estimate}
\usage{
f_geom_line_endpoint(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{all the geom_point arguments}
}
\value{

}
\description{
A geom that draws a point at the intercept of linear model estimate
}
\examples{
# without the function
library(magrittr)
library(dplyr)
cars \%>\%
  mutate(speed_cat = speed > 8) \%>\%
  group_by(speed_cat) \%>\%
  filter(speed == max(speed)) ->
 cars_group_endpoint

library(ggplot2)
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 aes(color = speed > 8) +
 geom_line() +
 geom_point(data = cars_group_endpoint)
layer_data(last_plot(), 2)

# the proto
"StatEndpoint"
f_geom_line_endpoint

# the function
f_geom_line_endpoint

# using the function
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 geom_line() +
 f_geom_line_endpoint(size = 4) +
 aes(color = speed > 9)
 layer_data(last_plot(), 2)

}
