% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_geom_prop_in_ts.R
\name{geom_tile_prop_over_time}
\alias{geom_tile_prop_over_time}
\title{Tiles the are filled according to proportion where total in x is denominator}
\usage{
geom_tile_prop_over_time(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Tiles the are filled according to proportion where total in x is denominator
}
\examples{
library(tidyverse)
gapminder::gapminder \%>\%
  mutate(gdp = gdpPercap * pop) ->
my_gapminder

my_gapminder \%>\%
  group_by(year, continent) \%>\%
  summarise(gdp = sum(gdp)) \%>\%
  mutate(prop_gdp = gdp/sum(gdp)) \%>\%
  ggplot() +
  aes(x = year) +
  aes(y = continent) +
  geom_tile() +
  aes(fill = prop_gdp) +
  scale_fill_viridis_c()

library(ggplot2)
library(magrittr)
my_gapminder \%>\%
ggplot() +
  aes(x = year) +
  aes(y = continent) +
  geom_tile_prop_over_time() +
  aes(fill = gdp) +
  scale_fill_viridis_c()
}
