% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_geom_c_hull_hollow.R
\name{k_geom_c_hull_hollow}
\alias{k_geom_c_hull_hollow}
\title{Percentage labels the are filled according to proportion where total in x is denominator}
\usage{
k_geom_c_hull_hollow(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Percentage labels the are filled according to proportion where total in x is denominator
}
\examples{
library(ggplot2)
library(magrittr)

chull(cars$speed, cars$dist) \%>\% # index of rim points
cars[.,] ->
cars_c_hull_rows

ggplot(data = cars) +
  aes(x = speed) +
  aes(y = dist) +
  geom_point() +
  geom_polygon(data = cars_c_hull_rows,
  alpha = .5
  )

ggplot(data = cars) +
  aes(x = speed) +
  aes(y = dist) +
  geom_point() +
  k_geom_c_hull_hollow(alpha = .5) +
  aes(color = speed >= 10)
}
