% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_geom_brick.R
\name{l_geom_brick}
\alias{l_geom_brick}
\title{Title}
\usage{
l_geom_brick(
  mapping = NULL,
  data = NULL,
  position = "stack",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
library(dplyr)
mtcars \%>\%
mutate(heavy = wt > 3) \%>\%
mutate(index = 1:n()) \%>\%
arrange(heavy) \%>\%
ggplot() +
  aes(x = factor(cyl)) +
  geom_bar(color = "grey85") +
  aes(group = index) +
  aes(fill = heavy)


mtcars \%>\%
mutate(heavy = wt > 3) \%>\%
ggplot() +
  aes(x = factor(cyl)) +
  l_geom_brick() +
  aes(fill = heavy) +
  aes(group = heavy) +
  scale_fill_manual(values = c("tan1",
  "steelblue3"))


}
