% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o_facet_sample_prop.R
\name{o_facet_sample_prop}
\alias{o_facet_sample_prop}
\title{Title}
\usage{
o_facet_sample_prop(
  n = 9,
  prop = 0.2,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  shrink = TRUE,
  strip.position = "top",
  seed = sample(2000:3000, 1)
)
}
\arguments{
\item{strip.position}{}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
set.seed(1323)
ggplot(data = cars) +
  aes(x = speed) +
  aes(y = dist) +
  geom_point(color = "olivedrab4") +
  c_geom_lm() +
  o_facet_sample_prop() +
  labs(title = "Random Sample, 20\% of population")
}
