% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bdts.R
\name{plot.bdts}
\alias{plot.bdts}
\title{Plot method for `<bdts>`}
\usage{
\method{plot}{bdts}(x, ...)
}
\arguments{
\item{x}{An `<bdts>` object}

\item{...}{[dots] not used and will produce a warning if extra
arguments are supplied}
}
\value{
A plot, called for side-effects.
}
\description{
Plot method for `<bdts>`
}
