% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_geom_point_sample.R
\name{q_geom_point_sample}
\alias{q_geom_point_sample}
\title{A geom that shows a sampling of points}
\usage{
q_geom_point_sample(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{all the geom_point arguments}
}
\value{

}
\description{
A geom that shows a sampling of points
}
\examples{
# without the function
library(magrittr)
library(dplyr)
set.seed(1214)
cars \%>\%
 slice_sample( n = 10) ->
 cars_sample


library(ggplot2)
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 geom_point() +
 geom_point(data = cars_sample,
            color = "magenta")

# the proto and function
"StatSamplepoint"
q_geom_point_sample

# using the function
ggplot(data = cars) +
 aes(x = speed, y = dist) +
 geom_point() +
 q_geom_point_sample(color = "red") +
 q_geom_point_sample(color = "green", sample_size = 5) +
 q_geom_point_sample(sample_size = 2, color = "blue")

}
