% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate a time series of data under an arbitrary simple generating process}
\usage{
sim_data(t, b, d)
}
\arguments{
\item{t}{Total time of the simulation.}

\item{b}{Birth rate.}

\item{d}{Death rate.}
}
\value{
<data.frame> with two numeric columns
}
\description{
Simulate a time series of data under an arbitrary simple generating process
}
