% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_theme_void_fill.R
\name{theme_void_fill}
\alias{theme_void_fill}
\title{Title}
\usage{
theme_void_fill(fill = "linen", color = fill)
}
\arguments{
\item{color}{}
}
\value{

}
\description{
Title
}
\examples{
ggplot(cars) +
  aes(x = speed) +
  aes(y = dist) +
  geom_point(color = "plum4") +
  theme_void() +
  theme(plot.background = ggplot2::element_rect(fill = "plum1", color = "plum1"),
        panel.background = ggplot2::element_rect(fill = "plum1", color = "plum1"))

ggplot(cars) +
  aes(x = speed) +
  aes(y = dist) +
  geom_point(color = "plum4") +
  theme_void_fill(fill = "plum1")
}
